Decouples Filesystem implementation
This refactors the way that the parser handles paths in general,
removing the filesystem-specific implementation details from the parsing
fuctions, and providing a new interface centered around providing the
correct readers through the `FileProvider` trait. Default
implementations for live-system and logarchive formats are provided, as
are utilities for categorizing files by path. This requird updates
throughout the repo's tests, and if merged, will require updates to the
larger tests that are currently shipping in the test archive.